pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "aim2"
version = "0.0.0"
description = "AIM2 toolkit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pydantic>=2.0", "PyYAML>=6.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88

[tool.isort]
profile = "black"
[project.scripts]
aim2-ontology = "aim2.cli.ontology:main"
